Test: Test functionality of thread_create() with full implementation.

+++ threads/thread.c
@@ -111,10 +111,10 @@ thread_create(void (*fn) (void *), void *parg)
 
     // Define the parameters of the new thread
     thread_queue[new_id].state = READY;
-    thread_queue[new_id].context->uc_mcontext.gregs[REG_RIP] = &thread_stub; // Instruction pointer for new thread should point to the stub function 
-    thread_queue[new_id].context->uc_mcontext.gregs[REG_RSP] = thread_queue[new_id].context->uc_stack.ss_sp; // Stack pointer should point to the end of the stack, stack grows downwards
-    thread_queue[new_id].context->uc_mcontext.gregs[REG_RDI] = fn; // Arg #1 for stub function
-    thread_queue[new_id].context->uc_mcontext.gregs[REG_RSI] = parg; // Arg #2 for stub function
+    thread_queue[new_id].context->uc_mcontext.gregs[REG_RIP] = (long long int)&thread_stub; // Instruction pointer for new thread should point to the stub function 
+    thread_queue[new_id].context->uc_mcontext.gregs[REG_RSP] = (long long int)thread_queue[new_id].context->uc_stack.ss_sp; // Stack pointer should point to the end of the stack, stack grows downwards
+    thread_queue[new_id].context->uc_mcontext.gregs[REG_RDI] = (long long int)fn; // Arg #1 for stub function
+    thread_queue[new_id].context->uc_mcontext.gregs[REG_RSI] = (long long int)parg; // Arg #2 for stub function
 
 	return THREAD_FAILED;
 }
